+++ ui/tests/unit/join-step-form.spec.ts
@@ -42,8 +42,13 @@ describe('join Step Form', () => {
       const wrapper = runner.shallowMount(initialState);
       const widgetMultiselect = wrapper.find('autocompletewidget-stub');
       expect(widgetMultiselect.props('options')).toEqual([
-        { trackBy: { type: 'ref', uid: '1' }, label: 'dataset1', $isDisabled: false },
-        { trackBy: { type: 'ref', uid: '2' }, label: 'dataset2', $isDisabled: true },
+        { trackBy: { type: 'ref', uid: '1' }, label: 'dataset1' },
+        {
+          trackBy: { type: 'ref', uid: '2' },
+          label: 'dataset2',
+          disabled: true,
+          tooltip: 'This dataset cannot be combined with the actual one',
+        },
       ]);
       expect(widgetMultiselect.props('withExample')).toEqual(true);
       expect(widgetMultiselect.props('trackBy')).toEqual('trackBy');